# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Stock Investment Assistant API
  version: "1.4.0"
  description: >
    REST API for the Stock Investment Assistant backend. Provides chat, configuration, and model catalog endpoints.
    
    
    **Architecture**: Built with Flask blueprints for modular organization:
    - Core API endpoints (health, chat, config) 
    - Model management endpoints (OpenAI model operations)
    - Socket.IO integration for real-time communication
    - Conversation-aware short-term memory for multi-turn conversations
    
    
    **Changes (v1.4.0)**:
    - Full API surface reconciliation (T030): openapi.yaml now documents all 26 registered endpoints
    - Migrated chat identity field from `session_id` to `conversation_id` to align with runtime code
    - Removed `session_id` field from ChatRequest/ChatResponse schemas (not implemented in code)
    - Removed ghost endpoint GET /api/sessions/{session_id}/conversation (never implemented)
    - Removed orphaned schemas: ConversationHistory, ConversationMessage, SessionArchivedError
    - Fixed 409 Conflict error schema to match actual code behavior (`CONVERSATION_ARCHIVED` code)
    - Added Phase C management API endpoints: workspaces (5), sessions (5), conversations (5)
    - Added missing session close endpoint (POST /api/sessions/{session_id}/close)
    - Added conversation summary endpoint (GET /api/conversations/{conversation_id}/summary)
...
    - Added context_overrides to conversation create request schema
...
    
...
servers:
...
paths:
...
  /api/workspaces/{workspace_id}/sessions:
    post:
      tags: [session-management]
      summary: Create a new session in a workspace
      description: >
        Creates a session within the specified workspace. The session starts in `active` status.
        The parent workspace must exist and be owned by the authenticated user.
      operationId: createSession
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/XUserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCreateRequest'
            examples:
              withTitle:
                value:
                  title: "NVDA earnings prep"
                  description: "Assess post-earnings risk/reward"
              minimal:
                value:
                  title: ""
      responses:
        "201":
          description: Session created
          headers:
            Location:
              description: Canonical URL for the created session
              schema:
                type: string
                example: "/api/sessions/550e8400-e29b-41d4-a716-446655440000"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
              examples:
                created:
                  value:
                    session_id: "550e8400-e29b-41d4-a716-446655440000"
                    workspace_id: "abc123"
                    title: "NVDA earnings prep"
                    description: "Assess post-earnings risk/reward"
                    status: "active"
                    conversation_count: 0
                    created_at: "2026-03-23T10:00:00Z"
                    updated_at: "2026-03-23T10:00:00Z"
        "400":
          description: Validation error (invalid request body)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                invalidBody:
                  value:
                    error:
                      message: "Request body must be a JSON object"
                      code: "VALIDATION_ERROR"
        "403":
          description: Ownership violation — workspace belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                forbidden:
                  value:
                    error:
                      message: "Access denied"
                      code: "FORBIDDEN"
        "404":
          description: Parent workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                notFound:
                  value:
                    error:
                      message: "Workspace not found"
                      code: "NOT_FOUND"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
    get:
      tags: [session-management]
      summary: List sessions in a workspace
      description: >
        Returns a paginated list of sessions within the specified workspace,
        sorted by `updated_at` descending. Only the workspace owner can list sessions.
      operationId: listSessions
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/XUserIdHeader'
        - name: limit
          in: query
          required: false
          description: Maximum number of items to return (1–100).
          schema:
            type: integer
            default: 25
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          description: Number of items to skip for pagination.
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: status
          in: query
          required: false
          description: Filter sessions by status.
          schema:
            type: string
            enum: [active, closed, archived]
      responses:
        "200":
          description: Paginated list of sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionList'
              examples:
                twoSessions:
                  value:
                    items:
                      - session_id: "550e8400-e29b-41d4-a716-446655440000"
                        workspace_id: "abc123"
                        title: "NVDA earnings prep"
                        description: null
                        status: "active"
                        conversation_count: 2
                        created_at: "2026-03-22T09:00:00Z"
                        updated_at: "2026-03-23T10:00:00Z"
                      - session_id: "660e8400-e29b-41d4-a716-446655440001"
                        workspace_id: "abc123"
                        title: "AMD analysis"
                        description: "Quarterly review"
                        status: "active"
                        conversation_count: 1
                        created_at: "2026-03-21T14:00:00Z"
                        updated_at: "2026-03-22T08:30:00Z"
                    total: 2
                    limit: 25
                    offset: 0
        "403":
          description: Ownership violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                forbidden:
                  value:
                    error:
                      message: "Access denied"
                      code: "FORBIDDEN"
        "404":
          description: Parent workspace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                notFound:
                  value:
                    error:
                      message: "Workspace not found"
                      code: "NOT_FOUND"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
  /api/sessions/{session_id}:
    get:
      tags: [session-management]
      summary: Get session detail
      description: >
        Returns the session with its conversation_count.
        Only the owning user (via workspace ownership) can access it.
      operationId: getSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/XUserIdHeader'
      responses:
        "200":
          description: Session detail with conversation count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
              examples:
                detail:
                  value:
                    session_id: "550e8400-e29b-41d4-a716-446655440000"
                    workspace_id: "abc123"
                    title: "NVDA earnings prep"
                    description: "Assess post-earnings risk/reward"
                    status: "active"
                    conversation_count: 3
                    created_at: "2026-03-22T09:00:00Z"
                    updated_at: "2026-03-23T10:00:00Z"
        "403":
          description: Ownership violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                forbidden:
                  value:
                    error:
                      message: "Access denied"
                      code: "FORBIDDEN"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                notFound:
                  value:
                    error:
                      message: "Session not found"
                      code: "NOT_FOUND"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
    put:
      tags: [session-management]
      summary: Update a session
      description: >
        Updates session `title` and/or `description`. Only the owning user
        can perform this operation.
      operationId: updateSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/XUserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionUpdateRequest'
            examples:
              updateTitle:
                value:
                  title: "NVDA post-earnings review"
              updateBoth:
                value:
                  title: "NVDA post-earnings review"
                  description: "Updated risk assessment after Q4 results"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
              examples:
                updated:
                  value:
                    session_id: "550e8400-e29b-41d4-a716-446655440000"
                    workspace_id: "abc123"
                    title: "NVDA post-earnings review"
                    description: "Updated risk assessment after Q4 results"
                    status: "active"
                    conversation_count: 3
                    created_at: "2026-03-22T09:00:00Z"
                    updated_at: "2026-03-23T11:00:00Z"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                invalidBody:
                  value:
                    error:
                      message: "Request body must be a JSON object"
                      code: "VALIDATION_ERROR"
        "403":
          description: Ownership violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                forbidden:
                  value:
                    error:
                      message: "Access denied"
                      code: "FORBIDDEN"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                notFound:
                  value:
                    error:
                      message: "Session not found"
                      code: "NOT_FOUND"
        "409":
          description: Conflict — stale entity (concurrent update detected)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                conflict:
                  value:
                    error:
                      message: "Session was modified by another request"
                      code: "CONFLICT"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
  /api/sessions/{session_id}/close:
    post:
      tags: [session-management]
      summary: Close a session
      description: >
        Transitions the session from `active` to `closed`. A closed session does not accept
        new conversations, but existing non-archived conversations remain message-capable.
        Only the owning user can close it.
      operationId: closeSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/XUserIdHeader'
      responses:
        "200":
          description: Session closed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
              examples:
                closed:
                  value:
                    session_id: "550e8400-e29b-41d4-a716-446655440000"
                    workspace_id: "abc123"
                    title: "NVDA earnings prep"
                    description: "Assess post-earnings risk/reward"
                    status: "closed"
                    conversation_count: 3
                    created_at: "2026-03-22T09:00:00Z"
                    updated_at: "2026-03-23T12:00:00Z"
        "403":
          description: Ownership violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                forbidden:
                  value:
                    error:
                      message: "Access denied"
                      code: "FORBIDDEN"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                notFound:
                  value:
                    error:
                      message: "Session not found"
                      code: "NOT_FOUND"
        "400":
          description: Validation error (e.g., session is not active)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                validationError:
                  value:
                    error:
                      message: "Invalid state transition"
                      code: "VALIDATION_ERROR"
        "409":
          description: Conflict — stale entity or concurrent modification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                conflict:
                  value:
                    error:
                      message: "Session was modified by another request"
                      code: "CONFLICT"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
  /api/sessions/{session_id}/archive:
    post:
      tags: [session-management]
      summary: Archive a session
      description: >
        Archives the session and all descendant conversations.
        Repeated archive requests are idempotent.
        Only the owning user can archive it.
      operationId: archiveSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - $ref: '#/components/parameters/XUserIdHeader'
      responses:
        "200":
          description: Session archived successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
              examples:
                archived:
                  value:
                    session_id: "550e8400-e29b-41d4-a716-446655440000"
                    workspace_id: "abc123"
                    title: "NVDA earnings prep"
                    description: "Assess post-earnings risk/reward"
                    status: "archived"
                    conversation_count: 3
                    created_at: "2026-03-22T09:00:00Z"
                    updated_at: "2026-03-23T12:00:00Z"
        "403":
          description: Ownership violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                forbidden:
                  value:
                    error:
                      message: "Access denied"
                      code: "FORBIDDEN"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                notFound:
                  value:
                    error:
                      message: "Session not found"
                      code: "NOT_FOUND"
        "400":
          description: Validation error (e.g., invalid lifecycle transition)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                validationError:
                  value:
                    error:
                      message: "Invalid state transition"
                      code: "VALIDATION_ERROR"
        "409":
          description: Conflict — stale entity or concurrent modification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
              examples:
                conflict:
                  value:
                    error:
                      message: "Session was modified by another request"
                      code: "CONFLICT"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagementError'
...
components:
  parameters:
...
    XUserIdHeader:
      name: X-User-ID
      in: header
      required: true
      description: User identifier used for ownership scoping (Mongo ObjectId string).
      schema:
        type: string
      example: "507f1f77bcf86cd799439011"
    WorkspaceId:
      name: workspace_id
      in: path
      required: true
      description: Unique identifier of the workspace.
      schema:
        type: string
      example: "abc123"
    SessionId:
      name: session_id
      in: path
      required: true
      description: Unique identifier of the session (UUID v4).
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
...
  schemas:
...
    ManagementError:
      type: object
      description: >
        Error envelope used by Phase C management API endpoints. The `error` field
        is an object containing a human-readable `message`, a machine-readable `code`,
        and optional `details` and `correlation_id`.
      properties:
        error:
          type: object
          properties:
            message:
              type: string
              description: Human-readable error description.
            code:
              type: string
              description: Machine-readable error code (e.g., VALIDATION_ERROR, NOT_FOUND, FORBIDDEN, INTERNAL_ERROR, CONFLICT).
            details:
              type: object
              additionalProperties: true
              description: Optional additional detail about the error.
            correlation_id:
              type: string
              format: uuid
              description: Optional correlation identifier for tracing.
          required: [message]
...
    SessionCreateRequest:
      type: object
      description: Request body for creating a session.
      properties:
        title:
          type: string
          description: Optional session title.
        description:
          type: string
          nullable: true
          description: Optional description of the session.
        assumptions:
          type: string
          nullable: true
          description: Optional contextual assumptions for the session.
        pinned_intent:
          type: string
          nullable: true
          description: Optional pinned intent or goal for the session.
        focused_symbols:
          type: array
          items:
            type: string
          description: Optional list of ticker symbols this session focuses on.
    SessionUpdateRequest:
      type: object
      description: Request body for updating a session. At least one field should be provided.
      properties:
        title:
          type: string
          description: Updated session title.
        description:
          type: string
          nullable: true
          description: Updated description.
...
    SessionDetail:
      type: object
      description: Session with conversation count.
      allOf:
        - $ref: '#/components/schemas/Session'
        - type: object
          properties:
            conversation_count:
              type: integer
              minimum: 0
              description: Total number of conversations in this session.
    SessionListItem:
      type: object
      description: Session item in list responses, includes conversation_count and focused_symbols.
      allOf:
        - $ref: '#/components/schemas/Session'
        - type: object
          properties:
            conversation_count:
              type: integer
              minimum: 0
              description: Total number of conversations in this session.
            focused_symbols:
              type: array
              items:
                type: string
              description: Ticker symbols this session focuses on.
    SessionList:
      type: object
      description: Paginated list of sessions.
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SessionListItem'
        total:
          type: integer
          minimum: 0
          description: Total number of matching sessions.
        limit:
          type: integer
        offset:
          type: integer
      required: [items, total, limit, offset]
...
security: []